requirements.txt
google-genai
google-auth-oauthlib
google-api-python-client
google-auth-httplib2


python-dotenv
pydub
numpy
requests

elevenlabs

langgraph
langchain-google-genai

streamlit
audio_recorder_streamlit
streamlit-float

tiktoken

firebase-admin
